refactor(cart): extract CartFileIo (export/import launchers)
Pull the SAF register/launch pair + JSON export/import handlers out of
CartActivity into their own class. Filename helpers move to
util/CartShareFile.kt where they're colocated with the share-file
plumbing that already uses them.
